Add basic infdev-20100611 support
This commit adds *very* basic support for viewing the """biomes""" of
indev-20100611. "Biomes" meaning water vs. not water, as this version
was well before the introduction of biomes.

The Code is functional, however it is both incredibly slow (as we are
generating real chunks, not even in parallel) and not side-effect free
(due to needing a temporary directory). I belive these issues can be
lessened sigificantly by removing the need to create a World instance
for the chunk generator; this would likely require some bytecode
hacking.