Avoid instantiating World
This commit avoids creating a World instance in the infdev interface by
modifying the bytecode of the loaded Chunk class to not dereference its
world. As a result, it is possible to run the chunk generator without
a world.

Factory objects are now allowed to define a ClassLoaderFactory which
will be used to get the class loader to load the associated classes.
Existing interfaces simply use the same class loader that was used to
determine the RecognisedVersion. The bytecode handling is then performed
by this class loader, using the ASM library.

Creating a world required setting up a temporary directory and was
probably not thread-safe. I'm not 100% sure that this was worth adding a
dependency for...