Support early betas
This commit implements the functionality of the support_early_betas
branch from 2016. The code is mostly a translation into the current set
of classes / builders, though I needed to add a bit of infrastructure.

I didn't add oceans because having some biomes is better than none and I
do not want to expand the chunk-gen mess past a1.1.